requirements.txt
pandas>=2.2,<3.0
numpy>=2.0,<3.0
yfinance==0.2.54
SQLAlchemy>=2.0,<3.0
python-dotenv>=1.0
pytz>=2024.1
flask>=3.0,<4.0
gunicorn>=21.2
requests>=2.31
